smb: fix integer overflow in NTLM AUTHENTICATE LM/NT response bounds check
lm/nt_response_offset (uint32) + len (uint16) was evaluated in 32-bit math, so a crafted offset wraps past the bound and sets a wild pointer later dereferenced in NTLMv2 validation. Cast to size_t like the sibling EncryptedRandomSessionKey check. Pre-auth OOB read.

Fixes #956

+++ src/server/smb/smb_ntlm.c
@@ -793,7 +793,7 @@ validate_authenticate(
     memcpy(&lm_response_len, buf + 12, 2);
     memcpy(&lm_response_offset, buf + 16, 4);
 
-    if (lm_response_offset + lm_response_len > buf_len) {
+    if ((size_t) lm_response_offset + lm_response_len > buf_len) {
         smb_ntlm_error("NTLM: Invalid LM response field");
         goto cleanup;
     }
@@ -803,7 +803,7 @@ validate_authenticate(
     memcpy(&nt_response_len, buf + 20, 2);
     memcpy(&nt_response_offset, buf + 24, 4);
 
-    if (nt_response_offset + nt_response_len > buf_len) {
+    if ((size_t) nt_response_offset + nt_response_len > buf_len) {
         smb_ntlm_error("NTLM: Invalid NT response field");
         goto cleanup;
     }
